restructure javascript directory
