+++ project/frontend/js/constants.js
@@ -8,7 +8,8 @@ export const API_MY_PROFILE = BACKEND_URL + "i/uac/me/";
 
 
 // MOCKUP API
-export const API_FRIEND_LIST = "./js/mock/mockFriendList.json";
+export const MOCKUP_PROFILE = "./js/mock/mockProfile.json";
+export const MOCKUP_FRIENDLIST = "./js/mock/mockFriendList.json";
 
 export const API_PROFILE_BY_ID = "api/user/";
 // export const API_GET_USER_PROFILE = "../js/mock/mockProfiles.json";
