+++ project/frontend/js/userManage/main.js
@@ -1,6 +1,6 @@
-import * as constant from "./constants.js"
-import { fetchAPI } from "./api.js"
-import { loadPage } from "./router.js";
+import * as constant from "../constants.js"
+import { fetchAPI } from "../api.js"
+import { loadPage } from "../router.js";
 
 async function getProfile() {
   try {
