+++ project/frontend/js/userManage/register.js
